xinv3d (1.7.0), 2023 July 17, Added 2nd player mode to Xinvaders3D.
It started with spotting a variable that ended as _1 which suggests
the possibility of adding a _2. This turned into a project that was
larger than expected with major updates and modifications added.

Noticed that camera.h has unfinished hints of camera_1 and camera_2
but for now, minimum mods were only done for camera.c instead.

Possible future TODO would be to make use of the camera.h structure